Adapt to module reorganization
